First rudimentary bits for HTML template rendering
Mustache sucks as template engine - unfortunately, D template renderers
prefer the template to be compiled into the main application for speed,
and I want the templates to be external.
Consequence of that is that some hacks are now in place to teach
Mustache at least basic template inheritance. This should be enough for
this basic usecase.