Give the GC some hints
While usually hinting the GC is considered a bad idea, this is probably
one of the few places where calling it to do a full collection run is a
good idea.
While processing data, we allocate large chunks of metadata which are
linked to each other and can (and should) ideally be freed after
processing one architecture. We know that, the GC doesn't, which makes
it run too late on some systems, which gets us OOM-killed.
By running a collection manually at the appropriate times, we can
prevent this. Also, after a processing run is the perfect time for a
stop-the-world GC to run, since at that time, we do not have any active
threads except for the main one.