Another GC hint
While the application runs fine under memory pressure (the GC seems to
try harder to free memory when there is less RAM available), it fails to
spawn other processes, like "xz" which we need for compression.
This should mitigate the problem. Long-term we need a way to use less
memory, or simple run on a machine with >> 4G of RAM.
Valgrind didn't spot any bigger memory leaks.