Make Arch backend work
This is just very basic work on the backend, it needs a lot of
polishing.
Especially the ArchiveDecompressor needs a bit of work to support the
usecase of Arch better, since the thing that is slowing down the backend
the most is the initial "fetch all data" stage, where the .files.tar.gz
tarball is processed.