Ensure gzipped files have repeatable results
The gz format puts a timestamp in the compressed file. Therefore
repeated runs produce different gz files even though the input data
is the same. Use a timestamp of 0 to ensure the compressed data is
the same when the input data is the same.

Neither the gzip executable nor the libarchive library provide a way
to specify the timestamp explicitly, so we set the modification time
of the input file. This is going to be removed anyway, so it doesn't
matter what it's set to.