Make icon tarball build reproducible
This has some quirks still, for example for some reasons multiple gcids
are found for the same package-id. This should actually not be possible,
and it needs to investigated why this happens.

Also, the metadata is not affected by this. The reproducibility of
metadata will be solved by not rebuilding it at all unless something has
changed, which is the more efficient approach compared to papering over
the issue by making GZip comply.