Store media in pool subdirectory
This is a prerequisite for the planned immutable suites feature. Since
it is a quite invasive change (needing manual intervention to not break
things from the generator user), we better do this now than later.
Using the pool subdirectory ensures full forward compatibility with
future versions which have immutable suites.