Move all path config to the Config singleton
This also moves the previous "cache" into a new "db" directory, since
the databases are actually not a mere cache anymore (at least not the
main one, the contentscache really is one).
We can now use the cache directory for actual cache data, starting my
moving the temporary directory there.
The tmpdir can get quite big and if /tmp is on a tmpfs this can start to
become a problem quickly. With storing it in the cache, we have proper
control over the data and can also regularly clean it up if there are
leftover files.