Don't name the databases classes "cache"
While "cache" is technically correct for the ContentsCache, it's also
highly misleading, since the DataStore is used for more than just cache
data (e.g. it is storing statistics).
Also, removing the databases can lead to seriously degraded performance,
not calling them "cache" makes this a bit more obvious.
Also, those data stores live in the db/ subdirectory on the workspace,
not the cache/ subdirectory (which is for actual cache data and can be
removed), which makes it even more useful to rename the classes.