Only retrieve DDTP translations from the section we're looking at
In this case

  main/i18n/Translation-foo
  contrib/i18n/Translation-bar

we were attempting to open main/i18n/Translation-bar, which doesn't
exist. This was causing a code path to short circuit the function to
load translations, meaning that some were missed.

Fix that to 'continue' instead of return. Also only return translations
in the suite we are asking about, not all translations.