Only add localized package descriptions if component has no description
This will result in people either getting a translated description for
their language which matches what upstream provided, or getting the
fallback description.
This makes the code a bit simpler and might be more what upstream
projects and users expect.

I am not 100% sure yet that this is a good decision though, since it
means that there might be less translations...
On the other hand, we don't get completely different descriptions for
some languages, which might be very confusing.