Remove crash causes my what is likely a Fontconfig bug and no GC issue
Not sure why, but when we don't immediately use the fpattern variable we
received from Fontconfig, later calls on it smash memory which makes it
impossible for the GC to free the memory later (from FreeType, not
Fontconfig! - which makes no sense), but works when freeing it
explicitly.

I currently assume that these issues are all multithreading-related and
due to Fontconfig being absolutely non-threadsafe.
It doesn't really make sense for these weird bugs to occur, neither does
it make sense that this change fixes the issues. But since others use
Fontconfig the same way, and with these changes I didn't see any crash
in 30 runs, this patch seems to reliably fix the bug.