Experiment with scoped classes
See https://dlang.org/phobos/std_typecons.html#.scoped.scoped
This unfortunately is unsafe, so we need to ensure that the scoping is
always preserved (the compiler doesn't do that for us).

On the upside, this might give us some more deterministic memory
management.

This still needs to be tested properly with a bigger repo!