Use separate variables for values in the addContents transaction
This might help resolving an incredibly hard to track down bug where
contents are added to the contents subdb, but icons aren't added to the
icons subdb.
Since this is one LMDB transaction, it should not be possible to have it
only partially succeed. This patch is a shot in the dark, to maybe
resolve this problem.