trivial: Rename section_overview to section_page
This naming confused me multiple times already, and I named the files
that way in the first place... The difference between "index" and
"overview" isn't clear ("index" is an overview of the sections, while
"overview" was meant to be an overview of the issues and metainfo).

So, rename the file to be consistent in naming with the other templates.