Relax icon scaling rules
This will allow asgen to scale down any bigger icon found in a XDG icon
theme, even if the theme doesn't explicitly permit it (e.g. if an icon
is "Threshold"). This will allow more apps to get their proper icon, and
is also much easier to understand to people - at the moment, asgen will
throw an "icon not found" error, even if e.g. an 192x192 icon exists.

This is a small divergence from the spec, but worthwile for this
purpose.