debian: Basic in-memory deduplication of translated texts for packages
This is a very conservative approach to avoid storing the same text over
and over again for each architecture.
We use a package-name/version as key to find out if we already have a
localization. In theory we could just use suite/section/package-name as
key to avoid duplicating descriptions for multiple versions.