ubuntu: Share a language pack provider between packages
This ensures we keep only one copy of the language data around, and also
separates the language pack stuff from the rest of the package code.
In total, this should even be slightly more memory efficient than the
previous solutions, and ensure we never try to modify the language pack
data from multiple threads at the same time.