Unconditionally add stock-type icon if desktop-file allows for it
The previous behavior was to only add the icon as "stock" if we found it
somewhere in an icon theme.
That is insufficient though and leads to incomplete styling in siftware
centers which might have newer/different/custom icon themes set.
Additionally, the previous behavior was also buggy because it failed to
determine between "icon is in an icon theme" and "icon is just in a
separate package" cases.