Refactor locale statistics loader
Since locale can be in pretty much any installed package, more advanced
logic is needed to find them.
The current implementation is only for Gettext, but adding Qt
translations will be much easier now.
The new implementation also *requires* the presence of a <translation/>
tag in order for this feature to work. This will likely not be changed,
explicitness just beats the complexity required for heuristics here.

CC: #37