snap: Make the GIO modules be found
For libsoup to be able to do SSL it needs to be able to find the GIO
modules that live in the libdir under $SNAP. Set $GIO_EXTRA_MODULES so
that they can be discovered here.

They need to use GSettings schemas; make sure we've compiled these and
set $GSETTINGS_SCHEMA_DIR so that they're found too.