backends/debian/debpkg: Forget a package's filename when removing its temporary dir
When downloading remote packages, we place them into a package specific
temporary directory. This is so we can remove them once we think we're
done, so that peak space usage is minimised.

We decide if a package is downloaded or not by the 'local' filename
being set. If it's set, we return this path to callers and then they can
try to operate on it. The problem is, we were failing to clear this
variable when removing the temporary directory in some circumstances.
That was meaning that a stale path could be given out to callers, a path
to a file that doesn't exist any more.

Fix this by always clearing the local path when removing the temporary
directory. That means that the next time the file is requested it will
be re-downloaded if this is a remote path.