various: Improve builddir != srcdir builds and build the snap like this
Newer versions of snapcraft seem to have changed the cwd of the build
step. It's more robust to do this by absolute path anyway, so do that.

A couple of parts of the build system want fixing for that-

  - Run the tests from the root of the source tree, and
  - Find data assets relative to the cwd

Another alternative would be to do something like
`G_TEST_{BUILD,SRC}DIR` that GLib has - environment variables that can
be defined to help tests find their assets. But this works OK for now.

Also run the sedding on the .gir file in override-pull; if the build
step is run twice then this results in messed up files otherwise.