ci: Ignore dscanner tests for now
D-Scanner is crashing randomly on the Debian CI, either by trying to
access an emptystring filename and running into an error, or in dparser
with a segmentation fault. The issue is very hard to reproduce locally,
and after spending a few hours on it, I think it's better to temporarily
ignore this issue and move on.
(dparser has always been very unstable after all)