Fix compilation due to broken generated GdkPixbuf code
Gir-to-d tries to handle the return value of 'gdk_pixdata_from_pixbuf'
as if it is zero-terminated.
Which causes it to try to use 'getArrayLength' on a void*, causing a
compilation error.
Temporary fix for now is to just not export that function.

Additionally, Add missing include for 'FILE*' that GdkPixbufModule needs
inside 'girepo/gdkpixbuf/c/types.d'.

Signed-off-by: Dan Printzell <xwildn00bx@gmail.com>