config: Move setUseOptipng call after we determine if optipng is available
Currently we tell AsCompose to use optipng just based on the value of
the feature flag. AsCompose will respect what we tell it without further
checking. If optipng is in fact not available then we'll crash later on
when we first try to use it.

Fix this by setting the flag *after* we check if optipng is installed.