snapcraft: stage libzstd1
appstream-generator depends on libarchive, which depends on libzstd1,
however it appears to not be staged into the snap. This results in
libzstd1 being used from the host system which may or may not be
compatible with the snap staged libarchive library. In order to
produce consistent results stage libzstd1 into the snap as well.