Initial implementation of data extractor using the compose API
There are still remaining todo items, for example the injected data
should use the same codepath as the regularly generated data (going
through AscCompose if possible), there is some optimization potential by
reusing the data extractor in the same thread, the screenshot handler
filesize limit needs to be added to ascompose as well as a bunch of
settings have to be revised (most importantly the icon settings).

Besides that, this code is usable but will need a lot of testing before
we can release it.