D-lete delicious D delights, damn
D is a fantastic language, and a joy to work with.
Sadly, its ecosystem is very small, and while maintaining
appstream-generator, me and others spend a lot of time making the D
ecosystem itself work, adding Meson integration, ensuring gir-to-d does
the right thing, fixing odd memory issues, fighting compiler bugs, etc.

With every compiler release something breaks, or features are
deprecated/removed from the standard library, which is annoying. Plus,
some people would like to use asgen in more scenarios (e.g. as a
library, potentially), and D's GC interferes with that.

C++23 has gained most of D's features - it has all the legacy C++,
and template programming is so much easier in D, but still, moving from
D -> C++ is actually fairly simple now. It moves asgen to a much more
established ecosystem, and allows its use in places where D was an
obstacle before.

So, while I generally absolutely hate rewrites / programming language
translations, for the future health of this project and to enable its
use in more organizations, switching it to C++ seems like the better
choice rather than to continue with D at this point.

Many thanks to the fantastic D community - the language itself is great,
despite IMHO needing a better future vision at this point.
D remains the "C++ without most of the bad stuff", and the seamless
integration with C codebases could be amazing if some of the friction
points were removed. I may still use the language in future.

For now though, asgen benefits from C++ in many ways, and I think using
it will be the right decision for the future of this particular project.