Allow backends to define a custom, non-default data prefix
This is needed for e.g. FreeBSD. Data in /usr locations will still be
processed, so icon-theme information and related necessary files will
still be found even if a custom prefix is set.

Currently, the prefix is backend-defined, but if we need to, we can make
it a config option instead at a later time.

Resolves: #132